+++ QuantumRefrigiz/DrawSoldireQ.cs
@@ -30,7 +30,7 @@ namespace QuantumRefrigiz
         public bool OnlySelfT = false;
         public bool AStarGreedyHuristicT = false;
         public bool ArrangmentsChanged = false;
-        public static double MaxHuristicxS = Double.MinValue;
+        public static int MaxHuristicxS = int.MinValue;
         public float RowS, ColumnS;
         public Color color;
         public ThinkingQuantumChess[] SoldierThinkingQuantum = new ThinkingQuantumChess[AllDraw.SodierMovments];
@@ -62,7 +62,7 @@ namespace QuantumRefrigiz
         public bool MaxFound(ref bool MaxNotFound)
         {
             
-                double a = ReturnHuristic();
+                int a = ReturnHuristic();
                 if (MaxHuristicxS < a)
                 {
                     Object O2 = new Object();
@@ -79,9 +79,9 @@ namespace QuantumRefrigiz
             MaxNotFound = true;
             return false;
         }
-        public double ReturnHuristic()
+        public int ReturnHuristic()
         {
-            double a = 0;
+            int a = 0;
             for (var ii = 0; ii < AllDraw.SodierMovments; ii++)
                 
                     a += SoldierThinkingQuantum[ii].ReturnHuristic(-1, -1, Order,false);
@@ -296,9 +296,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(S[0], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         if (LastRow != -1 && LastColumn != -1)
                                         {
@@ -348,9 +348,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(S[1], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         if (LastRow != -1 && LastColumn != -1)
                                         {
@@ -390,9 +390,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(DrawMinisterQ.M[0], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         if (LastRow != -1 && LastColumn != -1)
                                         {
@@ -424,9 +424,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(DrawMinisterQ.M[1], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         if (LastRow != -1 && LastColumn != -1)
                                         {
@@ -464,9 +464,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(DrawCastleQ.C[0], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         if (LastRow != -1 && LastColumn != -1)
                                         {
@@ -499,9 +499,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(DrawCastleQ.C[1], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         if (LastRow != -1 && LastColumn != -1)
                                         {
@@ -539,9 +539,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(DrawElefantQ.E[0], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         if (LastRow != -1 && LastColumn != -1)
                                         {
@@ -572,9 +572,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(DrawElefantQ.E[1], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         if (LastRow != -1 && LastColumn != -1)
                                         {
@@ -612,9 +612,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(DrawElefantQ.E[0], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         //if (Prob > 0)
                                         {
@@ -652,9 +652,9 @@ namespace QuantumRefrigiz
                                     g.DrawImage(DrawElefantQ.E[1], new Rectangle((int)(Row * (float)CellW), (int)(Column * (float)CellH), CellW, CellH));
                                     if (RingHalf)
                                     {
-                                        double Prob = 180;
+                                        int Prob = 180;
                                         if (AllDraw.Less != 0)
-                                            Prob = 180 * (AllDraw.Less / Double.MaxValue);
+                                            Prob = 180 * (AllDraw.Less / int.MaxValue);
                                         g.DrawArc(new Pen(new SolidBrush(Color.Red)), new Rectangle((int)((Row * (float)CellW)), (int)(Column * (float)CellH), CellW, CellH), -45, (int)Prob);
                                         //if (Prob > 0)
                                         {
