+++ RfregitzD/DrawElefant.cs
@@ -29,7 +29,7 @@ namespace RefrigtzDLL
         public bool OnlySelfT = false;
         public bool AStarGreedyHuristicT = false;
         public bool ArrangmentsChanged = false;
-        public static double MaxHuristicxE = -20000000000000000;
+        public static long MaxHuristicxE = -20000000000000000;
         public float Row, Column;
         public ThinkingChess[] ElefantThinking = new ThinkingChess[AllDraw.ElefantMovments];
         public int[,] Table = null;
@@ -56,7 +56,7 @@ namespace RefrigtzDLL
         public bool MaxFound(ref bool MaxNotFound)
         {
 
-            double a = ReturnHuristic();
+            int a = ReturnHuristic();
             if (MaxHuristicxE < a)
             {
                 Object O2 = new Object();
@@ -73,9 +73,9 @@ namespace RefrigtzDLL
             MaxNotFound = true;
             return false;
         }
-        public double ReturnHuristic()
+        public int ReturnHuristic()
         {
-            double a = 0;
+            int a = 0;
             for (var ii = 0; ii < AllDraw.ElefantMovments; ii++)
 
                 a += ElefantThinking[ii].ReturnHuristic(-1, -1, Order, false);
